Generalize fold in IntList
I copied this solution from the book (letter by letter) and am using 
this as an opportunity to study the solution.

I currently find this pretty confusing, the A in the type signature
for `f`, in particular. Moreover, I’m still trying to distill my
understanding of the val names in the type signature of `f` within the
method signature of `fold`.

I feel like I’m developing an intuition for what a “fold” is, but hard
to tell….

* Move specs for #double to not be within #product